削除の権限を分けます

+++ src/main/java/com/example/VisaAPI/controller/UserDeleteController.java
@@ -34,31 +34,38 @@ public class UserDeleteController {
 			DateTimeFormatter formatter = DateTimeFormatter.ofPattern("yyyy/MM/dd HH:mm:ss");
 			String formattedDateTime = currentDateTime.format(formatter);
 			List<UserDeleteModel>  user = userDeleteService.CheckDeleteByUsername(userDeleteModel);
-			if(user.size()!=0) {
-				userDeleteModel.setNote(formattedDateTime +"に"+ userDeleteModel.getLoginUsername()+"が削除しました");
-				int userRole = userDeleteService.DeleteByUsernameRole(userDeleteModel);
-				if(userRole == 1) {
-						userDeleteModel.setStatusRole("DELETED");
-							userDeleteModel.setStatus("SUCCESS");
-							int information = userDeleteService.DeleteByUsernameUser(userDeleteModel);
-								if(information == 1) {
-									userDeleteModel.setStatusInformation("DELETED");
-								}else {
-									userDeleteModel.setStatusInformation("在留カード情報がありません");
-								}
+			List<UserDeleteModel>  role = userDeleteService.CheckRoleLoginUser(userDeleteModel);
+			userDeleteModel.setRole(role.get(0).getRole());
+			if(role.get(0).getRole().equals("ADMIN") || userDeleteModel.getUsername().equals(userDeleteModel.getLoginUsername())) {
+				if(user.size()!=0) {
+					userDeleteModel.setNote(formattedDateTime +"に"+ userDeleteModel.getLoginUsername()+"が削除しました");
+					int userRole = userDeleteService.DeleteByUsernameRole(userDeleteModel);
+					if(userRole == 1) {
+							userDeleteModel.setStatusRole("DELETED");
+								userDeleteModel.setStatus("SUCCESS");
+								int information = userDeleteService.DeleteByUsernameUser(userDeleteModel);
+									if(information == 1) {
+										userDeleteModel.setStatusInformation("DELETED");
+									}else {
+										userDeleteModel.setStatusInformation("在留カード情報がありません");
+									}
+						}else {
+							userDeleteModel.setNote(formattedDateTime +"に"+ userDeleteModel.getLoginUsername()+"が削除されませんでした");
+							userDeleteModel.setStatusRole("ERROR");
+							userDeleteModel.setStatusInformation("ERROR");
+							userDeleteModel.setStatus("DEFEATED");
+						}
 					}else {
-						userDeleteModel.setNote(formattedDateTime +"に"+ userDeleteModel.getLoginUsername()+"が削除されませんでした");
 						userDeleteModel.setStatusRole("ERROR");
 						userDeleteModel.setStatusInformation("ERROR");
 						userDeleteModel.setStatus("DEFEATED");
+						userDeleteModel.setNote(formattedDateTime +"に"+ userDeleteModel.getUsername()+"が削除されたまたは存じません");
+						return ResponseEntity.ok(userDeleteModel);
 					}
-				}else {
-					userDeleteModel.setStatusRole("ERROR");
-					userDeleteModel.setStatusInformation("ERROR");
-					userDeleteModel.setStatus("DEFEATED");
-					userDeleteModel.setNote(formattedDateTime +"に"+ userDeleteModel.getUsername()+"が削除されたまたは存じません");
-					return ResponseEntity.ok(userDeleteModel);
-				}
+			}else {
+				userDeleteModel.setStatus("DEFEATED");
+				userDeleteModel.setNote(formattedDateTime +"に"+ "権限がないので、削除できませんでした");
+			}
 			return ResponseEntity.ok(userDeleteModel);
 		}
 }
